server/requirements.txt
annotated-types
anyio
attrs
bitarray
blinker
CacheControl
cachetools
certifi
cffi
charset-normalizer
click
colorama
cryptography
distro
firebase-admin
Flask
Flask-SQLAlchemy
flask-validate-json
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
idna
iniconfig
itsdangerous
Jinja2
jsonschema
jsonschema-specifications
load-dotenv
MarkupSafe
msgpack
numpy
packaging
pandas
pdfreader
pillow
pluggy
proto-plus
protobuf
pyasn1
pyasn1_modules
pycparser
pycryptodome
pydantic
pydantic_core
PyJWT
pyparsing
pypdf
PyPDF2
pypng
pytest
python-dateutil
python-dotenv
pytz
qrcode
referencing
requests
rpds-py
rsa
six
sniffio
SQLAlchemy
tqdm
typing_extensions
tzdata
uritemplate
urllib3
Werkzeug
psycopg2-binary